backend/order-service/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
drf-yasg==1.21.7
